Added Cargo.toml with project information
Renamed LICENSE.md to MIT-LICENSE.md
In code base made changes to server/src/main.rs
Imported std library to handle i/o, invoke a tcp listener and stream bytes
and start a thread when the client binds to the server.
There is also code to create a buffer to streaming tcp traffic and read bytes from
that buffer.
Added logic, if there are zero bytes on attempted connect the client closed the connection
if not write to the buffer and echo that back to the client
If there is an error reading the buffer, trap the error and print an error message on the
server side

BREAK: The code isn't cannot compile at this stage

FIX: N/A